fig_img_caption_extract/requirements.txt
opencv-python == 4.1.0.25
tqdm == 4.35.0
pdf2image == 1.12.1
detectron2 == 0.1.3
Pillow == 7.0.0
pdfminer.six == 20181108
easydict==1.9
numpy==1.18.1
